chore: Rename ITransactionRepository.d.ts
