+++ src/application/interfaces/ITransactionRepository.ts
