+++ src/application/services/TransactionService.ts
@@ -1,6 +1,6 @@
 import Transaction, { type TransactionType } from "../../domain/entities/Transaction.ts"
 import Money, { type Currency } from "../../domain/value-objects/Money.ts"
-import type { ITransactionRepository } from "../interfaces/ITransactionRepository.d.ts"
+import type { ITransactionRepository } from "../interfaces/ITransactionRepository.ts"
 import NotFoundError from "../errors/NotFoundError.ts"
 
 export type CreateTransactionDto = {
